Expand workspace: QUIC transport, HTTP/1.1+HTTP/3 stack, core refactor, reactor hardening
New crates:
- starfish-quic: QUIC transport (RFC 9000/9001/9002) — packet I/O, AEAD,
  header protection, loss detection, NewReno/Cubic congestion control,
  stream multiplexing, flow control. Self-rated maturity 2/5 (research);
  known DoS/conformance gaps tracked in starfish-quic/docs/bugs.md.
- starfish-db: scaffold for the transactional database (aspirational).

starfish-http:
- Replace the single http_client with a full h1/ + h3/ client and server.
- HTTP/1.1: incremental parser, chunked coding, connection pool.
- HTTP/3: QUIC-based client/server with QPACK; cargo-fuzz harness and
  h1/h3 integration + interop tests.

starfish-core:
- Replace wrappers/ and safe_* collection layers with MapCollection,
  MapEntry, and Pair abstractions.
- Add YFastTrie; rewrite SkipList/SkipTrie/SortedList internals
  (atomic_ptr, unpublished_node, 3-bit marked pointer).
- New map-collection test harness and stress tests.

starfish-reactor:
- 3-state lock protocol, io_uring queue configuration, IOCP/kqueue
  reworks, completion signaler, init barrier; reactor idle/lifecycle tests.

starfish-crossbeam / starfish-epoch / starfish-tls:
- Epoch-guarded collection refactor, epoch stress tests, benchmarks,
  profiling binaries; epoch reclamation and TLS client/server updates.

Project tooling and docs:
- CLAUDE.md, .claude/ commands + review lenses, .github CI workflow.
- docs/ (ARCHITECTURE, DESIGN-INDEX, proposals) and per-crate docs/
  with design notes and bug trackers.

+++ starfish-reactor/src/cooperative_io/io_windows/pooling_io_manager.rs
@@ -1,3 +1,9 @@
+//! Polling-based fallback I/O manager for Windows.
+//!
+//! Provides `PoolingIOManager`, which polls pending I/O futures for completion
+//! rather than relying on IOCP notifications. Used as a fallback when IOCP
+//! association is not available for a given handle.
+
 use std::{cell::RefCell, collections::VecDeque, io};
 
 use windows::Win32::System::IO::CancelIoEx;
@@ -18,7 +24,7 @@ use super::io_wait_future::IOWaitFuture;
 ///
 pub(crate) struct PoolingIOManager {
     pending_io_futures: RefCell<VecDeque<*const IOWaitFuture>>,
-    active_io: i32,
+    active_io: usize,
 }
 
 impl PoolingIOManager {
